refactor: move all cli related functionality into a cli dir
