+++ lib/cli/args.js
@@ -1,6 +1,6 @@
 import { DEFAULT_BASE_PATH } from 'appium-base-driver';
 import { parseSecurityFeatures, parseDefaultCaps, parseInstallTypes } from './parser-helpers';
-import { INSTALL_TYPES, DEFAULT_APPIUM_HOME } from './driver-config';
+import { INSTALL_TYPES, DEFAULT_APPIUM_HOME } from '../driver-config';
 
 // sharedArgs will be added to every subcommand
 const sharedArgs = [
