+++ lib/cli/driver.js
@@ -2,11 +2,11 @@
 
 import _ from 'lodash';
 import path from 'path';
-import { KNOWN_DRIVERS } from './drivers';
+import { KNOWN_DRIVERS } from '../drivers';
 import DriverConfig, { INSTALL_TYPE_NPM, INSTALL_TYPE_GIT, INSTALL_TYPE_GITHUB,
-                       INSTALL_TYPE_LOCAL } from './driver-config';
+                       INSTALL_TYPE_LOCAL } from '../driver-config';
 import NPM from './npm';
-import { errAndQuit, log, spinWith, JSON_SPACES } from './cli-utils';
+import { errAndQuit, log, spinWith, JSON_SPACES } from './utils';
 import { util, fs } from 'appium-support';
 
 const REQ_DRIVER_FIELDS = ['driverName', 'automationName', 'platformNames', 'mainClass'];
@@ -36,6 +36,8 @@ async function runDriverCommand (args) {
   if (json) {
     console.log(JSON.stringify(jsonResult, null, JSON_SPACES));
   }
+
+  return jsonResult;
 }
 
 class DriverCommand {
