+++ lib/cli/npm.js
