+++ lib/cli/parser-helpers.js
@@ -1,6 +1,6 @@
 import fs from 'fs';
 import _ from 'lodash';
-import { INSTALL_TYPES } from './driver-config';
+import { INSTALL_TYPES } from '../driver-config';
 
 // serverArgs will be added to the `server` (default) subcommand
 function parseSecurityFeatures (features) {
