+++ lib/cli/parser.js
@@ -2,8 +2,8 @@ import path from 'path';
 import _ from 'lodash';
 import { ArgumentParser } from 'argparse';
 import { sharedArgs, serverArgs, driverListArgs, driverInstallArgs, driverUninstallArgs,
-         driverUpdateArgs } from './cli-args';
-import { rootDir } from './utils';
+         driverUpdateArgs } from './args';
+import { rootDir } from '../utils';
 
 function getParser (debug = false) {
   const parser = new ArgumentParser({
