+++ lib/cli/utils.js
