+++ lib/main.js
@@ -6,7 +6,7 @@ import logger from './logger'; // logger needs to remain first of imports
 import _ from 'lodash';
 import { server as baseServer, routeConfiguringFunction as makeRouter } from 'appium-base-driver';
 import { asyncify } from 'asyncbox';
-import { default as getParser, getDefaultServerArgs } from './parser';
+import { default as getParser, getDefaultServerArgs } from './cli/parser';
 import { logger as logFactory, util } from 'appium-support';
 import {
   showConfig, checkNodeOk, validateServerArgs,
@@ -14,7 +14,7 @@ import {
   getGitRev, APPIUM_VER
 } from './config';
 import DriverConfig from './driver-config';
-import { runDriverCommand } from './driver-cli';
+import { runDriverCommand } from './cli/driver';
 import { AppiumDriver } from './appium';
 import registerNode from './grid-register';
 import { inspectObject } from './utils';
